refactor: add downloader module
