--- src-tauri/src/services/downloader.rs
@@ -1,265 +1,0 @@
-use crate::{inject_system_message, ExportMessage};
-use crate::protocol::types::{AppState, SystemLogLevel};
-use futures_util::StreamExt;
-use serde::{Deserialize, Serialize};
-use std::fs;
-use std::io::Write;
-use std::path::PathBuf;
-use chrono::{Local, TimeZone};
-use tauri::{AppHandle, Emitter, Manager};
-use tauri_plugin_shell::ShellExt;
-
-// --- 1. Define Constants for your GGUF Model ---
-pub const MODEL_FOLDER: &str = "Qwen3-Blue-Protocol-Translator-JA-KO";
-pub const MODEL_FILENAME: &str = "qwen3-4b-blueprotocol-ja2ko-q4_k_m.gguf";
-// Hugging Face direct download link (using /resolve/main/)
-pub const MODEL_URL: &str = "https://huggingface.co/enjay27/Qwen3-Blue-Protocol-Translator-JA-KO/resolve/main/qwen3-4b-blueprotocol-ja2ko-q4_k_m.gguf";
-pub const AI_SERVER_FOLDER: &str = "ai-server";
-pub const AI_SERVER_ZIP_URL: &str = "https://github.com/enjay27/resonance-stream/releases/download/v0.2.0/llama-b8157-bin-win-vulkan-x64.zip";
-pub const AI_SERVER_FILENAME: &str = "llama-server.exe";
-
-#[derive(Serialize, Clone)]
-pub struct FolderStatus {
-    pub exists: bool,
-    pub path: String,
-}
-
-#[derive(Serialize, Clone)]
-struct ProgressPayload {
-    pub current_file: String,
-    pub percent: u8,
-    pub total_percent: u8,
-}
-
-#[tauri::command]
-pub async fn check_model_status(app: tauri::AppHandle) -> Result<FolderStatus, String> {
-    // Check exactly one path for the .gguf file
-    let model_path = app.path()
-        .app_data_dir()
-        .map_err(|e| e.to_string())?
-        .join("models")
-        .join(MODEL_FOLDER)
-        .join(MODEL_FILENAME);
-
-    Ok(FolderStatus {
-        exists: model_path.exists(),
-        path: model_path.to_string_lossy().into_owned(),
-    })
-}
-
-#[tauri::command]
-pub async fn check_ai_server_status(app: tauri::AppHandle) -> Result<FolderStatus, String> {
-    // Check exactly one path for the .gguf file
-    let model_path = app.path()
-        .app_data_dir()
-        .map_err(|e| e.to_string())?
-        .join("bin")
-        .join(AI_SERVER_FOLDER)
-        .join(AI_SERVER_FILENAME);
-
-    Ok(FolderStatus {
-        exists: model_path.exists(),
-        path: model_path.to_string_lossy().into_owned(),
-    })
-}
-
-#[tauri::command]
-pub async fn download_model(app: AppHandle) -> Result<(), String> {
-    let model_dir = get_model_dir(&app)?;
-
-    fs::create_dir_all(&model_dir).map_err(|e| e.to_string())?;
-
-    let dest_path = model_dir.join(MODEL_FILENAME);
-
-    // Skip if already downloaded
-    if dest_path.exists() {
-        return Ok(());
-    }
-
-    let client = reqwest::Client::new();
-    let res = client.get(MODEL_URL).send().await.map_err(|e| e.to_string())?;
-
-    // ADD THIS CHECK: Ensure we actually got the file, not an HTML error page
-    if !res.status().is_success() {
-        return Err(format!("Failed to download model. Server returned: {}", res.status()));
-    }
-
-    let total_size = res.content_length().unwrap_or(0);
-
-    let mut file = fs::File::create(&dest_path).map_err(|e| e.to_string())?;
-    let mut downloaded: u64 = 0;
-    let mut stream = res.bytes_stream();
-
-    while let Some(item) = stream.next().await {
-        let chunk = item.map_err(|e| e.to_string())?;
-        file.write_all(&chunk).map_err(|e| e.to_string())?;
-        downloaded += chunk.len() as u64;
-
-        if total_size > 0 {
-            let percent = ((downloaded as f32 / total_size as f32) * 100.0) as u8;
-            let _ = app.emit("download-progress", ProgressPayload {
-                current_file: MODEL_FILENAME.to_string(),
-                percent,
-                total_percent: percent,
-            });
-        }
-    }
-
-    let _ = app.emit("download-progress", ProgressPayload {
-        current_file: "완료".into(),
-        percent: 100,
-        total_percent: 100,
-    });
-
-    Ok(())
-}
-
-fn get_model_dir(app: &AppHandle) -> Result<PathBuf, String> {
-    Ok(app.path()
-        .app_data_dir()
-        .map_err(|e| e.to_string())?
-        .join("models")
-        .join(MODEL_FOLDER))
-}
-
-pub fn get_model_path(app: &tauri::AppHandle) -> PathBuf {
-    app.path()
-        .app_data_dir()
-        .expect("Failed to resolve AppData directory")
-        .join("models")
-        .join(MODEL_FOLDER)
-        .join(MODEL_FILENAME)
-}
-
-#[tauri::command]
-pub async fn check_dict_update(app: tauri::AppHandle) -> Result<bool, String> {
-    let gist_url = "https://gist.githubusercontent.com/enjay27/487b588d38cd6bd514bc2be3d2db8270/raw/bp_dictionary.json";
-    let local_path = app.path().app_data_dir().unwrap().join("custom_dict.json");
-
-    // 1. Get the remote version (or last-modified header)
-    let client = reqwest::Client::new();
-    let response = client.get(gist_url).send().await.map_err(|e| e.to_string())?;
-    let remote_json: serde_json::Value = response.json().await.map_err(|e| e.to_string())?;
-
-    let remote_version = remote_json["version"].as_str().unwrap_or("0.0.0");
-
-    // 2. Read the local version
-    if let Ok(local_content) = std::fs::read_to_string(local_path) {
-        let local_json: serde_json::Value = serde_json::from_str(&local_content).unwrap_or_default();
-        let local_version = local_json["version"].as_str().unwrap_or("0.0.0");
-
-        // 3. Compare (e.g., "1.0.5" vs "1.0.4")
-        return Ok(remote_version != local_version);
-    }
-
-    Ok(true)
-}
-
-#[tauri::command]
-pub async fn sync_dictionary(app: tauri::AppHandle, state: tauri::State<'_, AppState>) -> Result<String, String> {
-    // 1. Define your Raw Gist URL
-    let url = "https://gist.githubusercontent.com/enjay27/487b588d38cd6bd514bc2be3d2db8270/raw/bp_dictionary.json";
-
-    // 2. Resolve Local Path: %APPDATA%/your.bundle.id/custom_dict.json
-    let dict_path = app.path().app_data_dir()
-        .map_err(|e| e.to_string())?
-        .join("custom_dict.json");
-
-    // 3. Fetch from Remote
-    let client = reqwest::Client::new();
-    let response = client.get(url).send().await.map_err(|e| e.to_string())?;
-    let json_content = response.text().await.map_err(|e| e.to_string())?;
-
-    // Validate JSON before saving
-    if serde_json::from_str::<serde_json::Value>(&json_content).is_err() {
-        return Err("Invalid JSON received from Gist".to_string());
-    }
-
-    // 4. Save Locally
-    std::fs::create_dir_all(dict_path.parent().unwrap()).map_err(|e| e.to_string())?;
-    std::fs::write(&dict_path, &json_content).map_err(|e| e.to_string())?;
-
-    inject_system_message(&app, SystemLogLevel::Success, "ModelManager", "Dictionary saved to AppData.");
-
-    inject_system_message(&app, SystemLogLevel::Success, "Translator", "Dictionary successfully synchronized.");
-
-    Ok("Dictionary updated and reloaded!".to_string())
-}
-
-// Add this command anywhere in downloader.rs
-#[tauri::command]
-pub async fn download_ai_server(app: AppHandle) -> Result<(), String> {
-    let ai_server_dir = app.path().app_data_dir().unwrap()
-        .join("bin")
-        .join(AI_SERVER_FOLDER);
-    fs::create_dir_all(&ai_server_dir).map_err(|e| e.to_string())?;
-
-    let server_exe = ai_server_dir.join("llama-server.exe");
-
-    // Skip if already downloaded and extracted
-    if server_exe.exists() {
-        return Ok(());
-    }
-
-    let zip_path = ai_server_dir.join("server_temp.zip");
-
-    // 1. Download the ZIP file (Streaming)
-    let client = reqwest::Client::new();
-    let res = client.get(AI_SERVER_ZIP_URL).send().await.map_err(|e| e.to_string())?;
-
-    // ADD THIS CHECK: Ensure we didn't hit a 404 on GitHub
-    if !res.status().is_success() {
-        return Err(format!("Failed to download AI server. Server returned: {}", res.status()));
-    }
-
-    let total_size = res.content_length().unwrap_or(0);
-
-    let mut file = fs::File::create(&zip_path).map_err(|e| e.to_string())?;
-    let mut downloaded: u64 = 0;
-    let mut stream = res.bytes_stream();
-
-    while let Some(item) = stream.next().await {
-        let chunk = item.map_err(|e| e.to_string())?;
-        file.write_all(&chunk).map_err(|e| e.to_string())?;
-        downloaded += chunk.len() as u64;
-
-        if total_size > 0 {
-            let percent = ((downloaded as f32 / total_size as f32) * 100.0) as u8;
-            let _ = app.emit("download-progress", ProgressPayload {
-                current_file: "AI 엔진 다운로드 중...".to_string(),
-                percent,
-                total_percent: percent,
-            });
-        }
-    }
-
-    // 2. Extract the ZIP file
-    let _ = app.emit("download-progress", ProgressPayload {
-        current_file: "압축 해제 중...".to_string(),
-        percent: 100,
-        total_percent: 100,
-    });
-
-    let zip_file = fs::File::open(&zip_path).map_err(|e| e.to_string())?;
-    let mut archive = zip::ZipArchive::new(zip_file).map_err(|e| e.to_string())?;
-
-    for i in 0..archive.len() {
-        let mut file = archive.by_index(i).map_err(|e| e.to_string())?;
-        let outpath = match file.enclosed_name() {
-            Some(path) => ai_server_dir.join(path.file_name().unwrap_or(path.as_os_str())), // Flattens the folder structure
-            None => continue,
-        };
-
-        if file.name().ends_with('/') {
-            continue; // Skip directories
-        }
-
-        let mut outfile = fs::File::create(&outpath).map_err(|e| e.to_string())?;
-        std::io::copy(&mut file, &mut outfile).map_err(|e| e.to_string())?;
-    }
-
-    // 3. Clean up the temp zip file
-    let _ = fs::remove_file(zip_path);
-
-    Ok(())
-}
